+++ src/Timestepping/diagnostics.jl
@@ -5,15 +5,15 @@ using ..Diagnostics: Diagnostics
 using ..Filaments: Filaments
 
 function Diagnostics.kinetic_energy_from_streamfunction(iter::VortexFilamentSolver; kws...)
-    (; ψs, fs, external_forcing, t,) = iter
+    (; ψs, fs, external_fields, t,) = iter
     Ls = BiotSavart.periods(iter.prob.p)
     E = Diagnostics.kinetic_energy_from_streamfunction(fs, ψs, iter.prob.p; kws...)
     # Add kinetic energy of external velocity field, if available.
     # Note that we only do this if we also included the streamfunction, since otherwise
     # we don't have enough information to estimate the total kinetic energy.
-    if external_forcing.velocity !== nothing && external_forcing.streamfunction !== nothing
+    if external_fields.velocity !== nothing && external_fields.streamfunction !== nothing
         E += Diagnostics.kinetic_energy_of_periodic_velocity_field(Ls) do x⃗
-            external_forcing.velocity(x⃗, t)
+            external_fields.velocity(x⃗, t)
         end
     end
     E