+++ test/runtests.jl
@@ -29,6 +29,8 @@ println()
 
 @testset "VortexPasta.jl" begin
     @includetest "plots.jl"
+    @includetest "synthetic_fields.jl"
+    @includetest "forcing.jl"
     @includetest "constants.jl"
     @includetest "vector_of_vector.jl"
     @includetest "padded_arrays.jl"
@@ -42,6 +44,7 @@ println()
     @includetest "ring_perturbed.jl"
     @includetest "ring_collision.jl"
     @includetest "ring_stretching.jl"
+    @includetest "ring_friction.jl"
     @includetest "background_vorticity.jl"
     @includetest "remove_small_filaments.jl"
     @includetest "trefoil.jl"